third_party/blink/web_tests/external/wpt/clipboard-apis/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] ClipboardChangeEvent interface object length
  assert_equals: wrong value for ClipboardChangeEvent.length expected 1 but got 0
[FAIL] ClipboardItem interface: attribute presentationStyle
  assert_true: The prototype object must have a property "presentationStyle" expected true got false
Harness: the test ran to completion.
